Update from internal repo

+++ s2e/qemu/s2e/Plugins/ExecutionTracers/TraceEntries.h
@@ -257,7 +257,7 @@ struct ExecutionTraceTestCase {
         ExecutionTraceTestCase *ret = (ExecutionTraceTestCase*)a;
 
         for(it = inputs.begin(); it != inputs.end(); ++it) {
-            Header hdr = {(*it).first.size(), (*it).second.size()};
+            Header hdr = {static_cast<uint32_t>((*it).first.size()), static_cast<uint32_t>((*it).second.size())};
             memcpy(a, &hdr, sizeof(hdr));
             a+=sizeof(hdr);
             memcpy(a, (*it).first.c_str(), (*it).first.size());
